+++ packages/shared/src/index.ts
@@ -6,3 +6,4 @@ export * from './error-codes';
 export * from './i18n-locale';
 export * from './slug-rules';
 export * from './git-provider';
+export * from './notification-template';